+++ NEWS.md
@@ -3,6 +3,10 @@ NEWS
 
 **TOSTER R package and jamovi module**
 
+# TOSTER v0.8.4
+- Added simple plot for `TOSTt` methods
+- Small fix to output for printed method for `TOSTt`
+
 # TOSTER v0.8.3
 - Change in the standard error formulation to Glass delta for independent samples
   - Hat tip to Paul Dudgeon for catching an error in the code that led to this development